@@ -11,7 +11,7 @@
   "scripts": {
     "start": "ng serve --proxy-config proxy.conf.json",
     "build": "ng build",
-    "build-prod": "ng build -prod",
+    "build-prod": "ng build --prod",
     "lint": "sass-lint -v --max-warnings 0 && ng lint --type-check",
     "lint-fix": "ng lint --type-check --fix",
     "lint-watch": "npm-watch lint",
@@ -21,32 +21,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.10",
-    "@angular/cdk": "5.2.5",
-    "@angular/common": "5.2.10",
-    "@angular/compiler": "5.2.10",
-    "@angular/core": "5.2.10",
-    "@angular/flex-layout": "5.0.0-beta.14",
-    "@angular/forms": "5.2.10",
-    "@angular/material": "5.2.5",
-    "@angular/platform-browser": "5.2.10",
-    "@angular/platform-browser-dynamic": "5.2.10",
-    "@angular/router": "5.2.10",
+    "@angular/animations": "6.0.0",
+    "@angular/cdk": "6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/flex-layout": "6.0.0-beta.15",
+    "@angular/forms": "6.0.0",
+    "@angular/material": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/router": "6.0.0",
     "core-js": "2.5.5",
     "hammerjs": "2.0.8",
-    "rxjs": "5.5.10",
+    "rxjs": "6.1.0",
     "web-animations-js": "2.3.1",
     "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.4",
-    "@angular/compiler-cli": "5.2.10",
-    "@types/node": "9.6.6",
+    "@angular-devkit/build-angular": "0.6.0",
+    "@angular/cli": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@types/node": "10.0.4",
     "codelyzer": "4.3.0",
     "jest": "22.4.3",
-    "jest-preset-angular": "5.2.1",
+    "jest-preset-angular": "5.2.2",
     "sass-lint": "1.12.1",
-    "tslint": "5.9.1",
-    "typescript": "2.5.3"
+    "tslint": "5.10.0",
+    "typescript": "2.7.2"
   }
 }
