@@ -19,12 +19,6 @@ import { PlayerStatsComponent } from './player-stats/player-stats.component';
 import { ShowAuthedDirective } from './shared/directives/show-authed.directive';
 import { DisplayErrorsComponent } from './shared/errors/display-errors.component';
 import { HeaderComponent } from './shared/header/header.component';
-import { ApiService } from './shared/services/api.service';
-import { AuthGuard } from './shared/services/auth-guard.service';
-import { EntriesService } from './shared/services/entries.service';
-import { NoAuthGuard } from './shared/services/no-auth-guard.service';
-import { TeamsService } from './shared/services/teams.service';
-import { UserService } from './shared/services/user.service';
 
 @NgModule({
   declarations: [
@@ -54,14 +48,7 @@ import { UserService } from './shared/services/user.service';
     MatToolbarModule,
     AppRoutingModule
   ],
-  providers: [
-    ApiService,
-    AuthGuard,
-    EntriesService,
-    NoAuthGuard,
-    TeamsService,
-    UserService
-  ],
+  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule {}
