@@ -3,7 +3,9 @@ import { CanActivate } from '@angular/router';
 
 import { UserService } from './user.service';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class NoAuthGuard implements CanActivate {
   constructor(
     private userService: UserService
